Move translator classes to c.g.z.c.j2se.translators
This eliminates the split package preventing use of this JAR as
a JPMS Module.
